+++ src/backend/external_api/engine.rs
@@ -45,6 +45,12 @@ impl std::fmt::Debug for Engine {
     }
 }
 
+impl Default for Engine {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
 impl Engine {
     pub fn new() -> Self {
         Self {